IMPLEMENTANDO CLIQUES NOS ITENS DO ESTOQUE, NO RECYCLERVIEW

+++ app/src/main/java/com/alphazer0/wmsgestaocomercial/ui/activities/clientes/ListaDeClientesActivity.java
@@ -22,6 +22,7 @@ import android.view.inputmethod.EditorInfo;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.SearchView;
+import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -31,6 +32,7 @@ import com.alphazer0.wmsgestaocomercial.R;
 import com.alphazer0.wmsgestaocomercial.database.ClientesDatabase;
 import com.alphazer0.wmsgestaocomercial.database.roomDAO.RoomClienteDAO;
 import com.alphazer0.wmsgestaocomercial.model.Cliente;
+import com.alphazer0.wmsgestaocomercial.model.Produto;
 import com.alphazer0.wmsgestaocomercial.ui.adapters.ListaClientesAdapter;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
@@ -61,10 +63,10 @@ public class ListaDeClientesActivity extends AppCompatActivity {
     private List<Cliente> clientes = new ArrayList<>();
     private final Context context = this;
     //CONFIGURACAO SCRIPT E PLANILHA BASE DADOS
-    String linkMacro= LINK_MACRO;
+    String linkMacro = LINK_MACRO;
     String idPlanilha = ID_PASTA;
-    private int put =0;
-    public int id =0;
+    private int put = 0;
+    public int id = 0;
 //==================================================================================================
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -83,47 +85,31 @@ public class ListaDeClientesActivity extends AppCompatActivity {
         adapter.atualiza(dao.todosClientes());
     }
 //==================================================================================================
-//    //MENU ITENS LISTA
-//    @Override
-//    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
-//    super.onCreateContextMenu(menu, v, menuInfo);
-//    getMenuInflater().inflate(R.menu.menu_listas_activity, menu);
-//}
-//
-//    @Override
-//    public boolean onContextItemSelected(MenuItem item) {
-//        int itemId = item.getItemId();
-//        if (itemId == R.id.menu_remover_listas_activity) {
-//            confirmaRemocao(item);
-//        }
-//        return super.onContextItemSelected(item);
-//    }
-//
-//    public void confirmaRemocao(final MenuItem item) {
-//    new AlertDialog.Builder(context).setTitle("Removendo Cliente").setMessage("Deseja mesmo remover o cliente?").setPositiveButton("Sim", (dialogInterface, i) -> {
-//        Cliente cliente = pegaCliente(item);
-//        put = 3;
-//        id = cliente.getId();
-//        new SendRequest().execute();
-//         remove(cliente);
-//    })
-//            .setNegativeButton("Não",null)
-//            .show();
-//    }
-//
-//
-//    private Cliente pegaCliente(MenuItem item) {
-//        AdapterView.AdapterContextMenuInfo menuInfo = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
-//        return adapter.getItem(menuInfo.position);
-//    }
-//==================================================================================================
-    private void remove(Cliente cliente){
-    adapter.remove(cliente);
-    dao.removeCliente(cliente);
+    //MENU ITENS LISTA
+    @Override
+    public boolean onContextItemSelected(MenuItem item) {
+        int itemId = item.getItemId();
+        if (itemId == 0) {
+            confirmaRemocao(item);
+        }
+        return super.onContextItemSelected(item);
+    }
+
+    public void confirmaRemocao(final MenuItem item) {
+        new AlertDialog.Builder(context).setTitle("Removendo Cliente").setMessage("Deseja mesmo remover o Cliente?").setPositiveButton("Sim", (dialogInterface, i) -> {
+            int position = item.getGroupId();
+            Cliente cliente = clientes.get(position);
+            Toast.makeText(context, "" + clientes.get(position), Toast.LENGTH_SHORT).show();
+            dao.removeCliente(cliente);
+            adapter.remove(position);
+            put = 3;
+        })
+                .setNegativeButton("Não", null)
+                .show();
     }
 //==================================================================================================
     private void configuraAdapter() {
-        adapter = new ListaClientesAdapter(this,clientes);
+        adapter = new ListaClientesAdapter(this, clientes);
         //PEGA TODOS OS CLIENTES DO BANCO DE DADOS
         dao = ClientesDatabase.getInstance(this).getClienteDAO();
     }
@@ -177,69 +163,71 @@ public class ListaDeClientesActivity extends AppCompatActivity {
     //CONFIGURACOES DO FILTRO PESQUISA
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-    MenuInflater inflater = getMenuInflater();
-    inflater.inflate(R.menu.menu_search, menu);
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.menu_search, menu);
 
-    MenuItem searchItem = menu.findItem(R.id.action_search);
-    SearchView campoBusca = (SearchView) searchItem.getActionView();
+        MenuItem searchItem = menu.findItem(R.id.action_search);
+        SearchView campoBusca = (SearchView) searchItem.getActionView();
 
-    campoBusca.setImeOptions(EditorInfo.IME_ACTION_SEARCH);
-    campoBusca.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-        @Override
-        public boolean onQueryTextSubmit(String query) {
-            return false;
-        }
-
-        @Override
-        public boolean onQueryTextChange(String newText) {
-            if (newText.toString() != null && !newText.toString().equals("")) {
-                adapter.getFilter().filter(newText);
-            }else {
-                adapter.atualiza(dao.todosClientes());
+        campoBusca.setImeOptions(EditorInfo.IME_ACTION_SEARCH);
+        campoBusca.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                return false;
             }
-            return false;
-        }
-    });
-    return true;
-   }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                if (newText.toString() != null && !newText.toString().equals("")) {
+                    adapter.getFilter().filter(newText);
+                } else {
+                    adapter.atualiza(dao.todosClientes());
+                }
+                return false;
+            }
+        });
+        return true;
+    }
 //==================================================================================================
     //INICIANDO COMUNICACAO WEB
     public class SendRequest extends AsyncTask<String, Void, String> {
 
-    protected void onPreExecute(){}
+        protected void onPreExecute() {
+        }
 
-    protected String doInBackground(String... arg0) {
-        try{
-            URL url = new URL(linkMacro);
-            String idPlan= idPlanilha;
+        protected String doInBackground(String... arg0) {
+            try {
+                URL url = new URL(linkMacro);
+                String idPlan = idPlanilha;
 
-            JSONObject enviaDados = enviaDadosParaPlanilha(idPlan);
-            HttpURLConnection connection = executaConeccaoExternalServer(url);
-            escreveDadosNaPlanilha(enviaDados, connection);
-            return verificaLinhaVazia(connection);
-        }//end try
+                JSONObject enviaDados = enviaDadosParaPlanilha(idPlan);
+                HttpURLConnection connection = executaConeccaoExternalServer(url);
+                escreveDadosNaPlanilha(enviaDados, connection);
+                return verificaLinhaVazia(connection);
+            }//end try
 
-        catch(Exception e){
-            return new String("Exception: " + e.getMessage());
-        }//end catch
+            catch (Exception e) {
+                return new String("Exception: " + e.getMessage());
+            }//end catch
 
-    }//end doInBackGround
+        }//end doInBackGround
 
-    @Override
-    protected void onPostExecute(String resultado) {
+        @Override
+        protected void onPostExecute(String resultado) {
 //            Toast.makeText(getApplicationContext(),"Salvo Na GSheet!!!",Toast.LENGTH_LONG).show();
 //            Toast.makeText(getApplicationContext(), resultado, Toast.LENGTH_LONG).show();
-    }//end onPostExecute
-}//end sendRequest
+        }//end onPostExecute
+    }//end sendRequest
+
 //==================================================================================================
     private String verificaLinhaVazia(HttpURLConnection connection) throws IOException {
         int codigoWeb = connection.getResponseCode();
         if (codigoWeb == HttpsURLConnection.HTTP_OK) {
             BufferedReader leValor = new BufferedReader(new InputStreamReader(connection.getInputStream()));
             StringBuffer sb = new StringBuffer("");
-            String linha="";
+            String linha = "";
 
-            while((linha = leValor.readLine()) != null) {
+            while ((linha = leValor.readLine()) != null) {
                 sb.append(linha);
                 break;
             }//end while
@@ -249,7 +237,7 @@ public class ListaDeClientesActivity extends AppCompatActivity {
 
         }//end if
         else {
-            return new String("false : "+codigoWeb);
+            return new String("false : " + codigoWeb);
         }//end else
     }
 
@@ -277,16 +265,16 @@ public class ListaDeClientesActivity extends AppCompatActivity {
     private JSONObject enviaDadosParaPlanilha(String action) throws JSONException {
         JSONObject enviaDados = new JSONObject();
 
-        if(put == 3){
-            action ="deleteCliente";
+        if (put == 3) {
+            action = "deleteCliente";
         }
 
-        enviaDados.put("pasta",ID_PASTA);
-        enviaDados.put("planilha",CLIENTE_PLAN);
-        enviaDados.put("action",action);
-        enviaDados.put("idCliente",id);
+        enviaDados.put("pasta", ID_PASTA);
+        enviaDados.put("planilha", CLIENTE_PLAN);
+        enviaDados.put("action", action);
+        enviaDados.put("idCliente", id);
 
-        Log.e("params",enviaDados.toString());
+        Log.e("params", enviaDados.toString());
         return enviaDados;
     }
 
@@ -296,13 +284,13 @@ public class ListaDeClientesActivity extends AppCompatActivity {
 
         Iterator<String> itr = params.keys();
 
-        while(itr.hasNext()){
-            String key= itr.next();
+        while (itr.hasNext()) {
+            String key = itr.next();
             Object value = params.get(key);
 
             if (first) {
                 first = false;
-            }else {
+            } else {
                 result.append("&");
             }
             result.append(URLEncoder.encode(key, "UTF-8"));
